+++ postmarker/__init__.py
@@ -1,4 +1,4 @@
 # coding: utf-8
 
 
-__version__ = '0.3.0.dev'
+__version__ = '0.3.0'