+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 
 - Improved attachments support. `#23`_
 - Improved MIME messages support. `#28`_
+- Jython support. `#13`_
 
 0.2.0 - 07.10.2016
 ------------------
@@ -36,5 +37,6 @@ Changelog
 .. _#23: https://github.com/FriendlyCoders/postmarker/issues/23
 .. _#18: https://github.com/FriendlyCoders/postmarker/issues/18
 .. _#16: https://github.com/FriendlyCoders/postmarker/issues/16
+.. _#13: https://github.com/FriendlyCoders/postmarker/issues/13
 .. _#12: https://github.com/FriendlyCoders/postmarker/issues/12
 .. _#11: https://github.com/FriendlyCoders/postmarker/issues/11